+++ packages/mobile/src/tabs/Tabs.tsx
@@ -32,6 +32,7 @@ const AnimatedBox = Animated.createAnimatedComponent(Box);
 type TabContainerProps = {
   id: string;
   registerRef: (tabId: string, ref: View) => void;
+  style?: StyleProp<ViewStyle>;
   children?: React.ReactNode;
 };
 
@@ -106,6 +107,8 @@ export type TabsProps<
     styles?: {
       /** Root container element */
       root?: StyleProp<ViewStyle>;
+      /** Container element wrapping each tab */
+      tabContainer?: StyleProp<ViewStyle>;
       /** Tab element */
       tab?: StyleProp<ViewStyle>;
       /** Active indicator element */
@@ -222,7 +225,12 @@ const TabsComponent = memo(
                 ...tabRest,
               };
               return (
-                <TabContainer key={id} id={id} registerRef={registerRef}>
+                <TabContainer
+                  key={id}
+                  id={id}
+                  registerRef={registerRef}
+                  style={styles?.tabContainer}
+                >
                   <RenderedTab {...renderedTabProps} />
                 </TabContainer>
               );