+++ packages/icons/CHANGELOG.md
@@ -8,6 +8,20 @@ All notable changes to this project will be documented in this file.
 
 <!-- template-start -->
 
+## 5.20.0 (6/24/2026 PST)
+
+#### 🚀 Updates
+
+- Feat: Publish icons 2026-06-24. [[#769](https://github.com/coinbase/cds/pull/769)]
+
+##### ⭐️ Added (5)
+
+- waveForm
+- record
+- stop
+- desktop
+- liveEvents
+
 ## 5.19.0 (6/9/2026 PST)
 
 #### 🚀 Updates